v0.2 r7
- Изменение погрешности теперь влияет на результат
- Исправлен алгоритм вычисления (В предыдущих версиях не были
задействованы некоторые переменные, что влияло на точность)

+++ PT/Data.cs
@@ -32,7 +32,7 @@ namespace PT
         /// <summary>
         /// Погрешность
         /// </summary>
-        public static double e;
+        public static int e;
 
         /// <summary>
         /// Искомый корень
